Updated Base Statemachine, Created Entity Class and moved PlayerComponents to a child of Entity called Player
Updated Base state manager to work with any kind of Entity (Player, enemy) by changing the constructor in child class.

Created an entity class with common methods inside to clean up code and make it reusable when creating the enemy.

Player is no longer a singleton and is instead passed to each state on construction.